+++ Engine/gui/pge_menuboxbase.h
@@ -64,7 +64,7 @@ class PGE_MenuBoxBase : public PGE_BoxBase
         void addMenuItems(std::vector<std::string>& menuitems);
 
         void setPos(double x, double y);
-        void setMaxMenuItems(int items);
+        void setMaxMenuItems(size_t items);
         void setBoxSize(double _Width, double _Height, double _padding);
         void update(double ticks);
         void render();
@@ -72,7 +72,7 @@ class PGE_MenuBoxBase : public PGE_BoxBase
         bool isRunning();
         void exec();
         void setRejectSnd(long sndRole);
-        int  answer();
+        size_t answer();
 
         void reject();
 
@@ -91,9 +91,9 @@ class PGE_MenuBoxBase : public PGE_BoxBase
         bool    running = false;
         int     fontID = 0;
         GlColor fontRgba;
-        int     _answer_id = 0;
+        size_t  _answer_id = 0;
 
-        long    reject_snd = 0;
+        size_t  reject_snd = 0;
         PGE_Point _pos;
         Controller *_ctrl1 = nullptr;
         Controller *_ctrl2 = nullptr;
